refactor: Fix redundant 'upgrade bytecode' and 'reinitialize bytecode' calls; consolidate; fix races
